backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-contrib-python-headless
pillow
joblib
scikit-learn
pydantic